requirements.txt
numpy>=1.19.0
torch>=1.7.0
tqdm>=4.60.0
matplotlib>=3.3.0
python-chess>=1.9.0
tensorboard>=2.4.0
pyyaml>=5.4.0 